added cjs and mjs support
