+++ fixup.sh
@@ -1,0 +1,10 @@
+cat >dist/cjs/package.json <<!EOF
+{
+    "type": "commonjs"
+}
+!EOF
+cat >dist/mjs/package.json <<!EOF
+{
+    "type": "module"
+}
+!EOF
