+++ package.json
@@ -3,20 +3,30 @@
   "version": "3.0.0",
   "description": "Instances of classes which are directly callable as functions.",
   "repository": "CGamesPlay/node-callable-instance",
-  "main": "dist/index.js",
-  "types": "src/types/index.d.ts",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/mjs/index.js",
+  "types": "lib/index.d.ts",
+  "exports": {
+    "import": "./dist/mjs/index.js",
+    "require": "./dist/cjs/index.js"
+  },
   "scripts": {
-    "build": "rm -rf ./dist && tsc",
-    "test": "mocha src/test && tsc"
+    "build": "rm -rf dist && tsc && tsc -p tsconfig-cjs.json && ./fixup.sh && npm link && npm link callable-instance",
+    "test": "npm run build && mocha src/test"
   },
   "keywords": [
     "instance",
     "function",
     "object",
     "class",
+    "constructor",
     "callable",
     "callable-instance",
-    "JS Callable Object"
+    "callable-object",
+    "callable-class",
+    "node-callable-instance",
+    "es6-callable",
+    "callable-constructor"
   ],
   "bugs": {
     "url": "https://github.com/CGamesPlay/node-callable-instance/issues"
@@ -24,6 +34,7 @@
   "author": "Ryan Patterson",
   "license": "MIT",
   "devDependencies": {
+    "@types/jest": "^29.5.5",
     "@types/mocha": "^10.0.1",
     "mocha": "^10.2.0",
     "prettier": "^2.8.1",
