+++ src/index.js
@@ -1,22 +1,21 @@
-export const CALL = Symbol("Callable.CALL")
-const BOUND = Symbol("Callable.BOUND")
-
+export const CALL = Symbol("Callable.CALL") // CALL symbol
+const BOUND = Symbol("Callable.BOUND") // unique symbol for binding function
 class Callable extends Function {
-  static get CALL() {
+  static get CALL() { // getter for CALL symbol
     return CALL
   }
-  get [Symbol.toStringTag]() {
+  get [Symbol.toStringTag]() { // creates stringTag (Object.prototype.toString.call(new Callable()) === [object Callable])
     return "Callable"
   }
   constructor(property = CALL) {
-    super("...a", "return this.a[this.b][this.c](...a)")
-    this[BOUND] = this.bind({
+    super("...a", "return this.a[this.b][this.c](...a)") // calls property from child class ( function (...args)){ return this[BOUND][property](...args)})
+    this[BOUND] = this.bind({ // binds to property, bound symbol and this
       a: this,
       b: BOUND,
       c: property
     })
-    Object.defineProperty(this[BOUND], "name", Object.getOwnPropertyDescriptor(this.constructor, "name"));
-    return this[BOUND]
+    Object.defineProperty(this[BOUND], "name", Object.getOwnPropertyDescriptor(this.constructor, "name")); // copies constructor name
+    return this[BOUND] // returns Bound function
   }
 }
